Major rewrite. Bumped version to 3.0.0. Config options have changed, see the readme.
Originally MFCLogger was written in CoffeeScript and later ported to
TypeScript via direct conversion. That resulted in TypeScript that even I
found hard to follow, and with a lot of implicit 'any' typings everywhere.
Plus the code was unnecessarily complicated, especially given the latest
features I've added to MFCAuto.

So this rewrite should make things more maintainable and it adds new
features.

* Enable specifying a when filter even when a specific model id is given.
* Make the viewers logger actually work, although it's only questionably
 useful

It's quite possible that I introduced some bugs with this change. So if
you have a working config and don't want to take advantage of the new
features right now, you might wait a week or two to let this change
stabilize.